requirements.txt
streamlit
pdfplumber
python-docx
PyMuPDF
openai
google-genai
